chore: fix clippy
